requirements.txt
PyMuPDF
pdf2image
paddleocr
python-docx
Pillow
numpy